COMP: Fixing warning: implicit conversion shortens 64-bit value into a 32-bit value on MacOS.

+++ Code/IO/itkDICOMImageIO2.cxx
@@ -82,7 +82,7 @@ void DICOMImageIO2::ReadDataCallback( doublebyte,
                                       unsigned char* val,
                                       quadbyte len)
 {
-  unsigned int imageBytes = this->GetImageSizeInBytes();
+  unsigned int imageBytes = static_cast<unsigned int>( this->GetImageSizeInBytes() );
   if (len < 0)
     {
     len = 0;
